+++ lib/recvmsg.c
@@ -6,13 +6,13 @@
 
 extern int socketcall(int callno,long* args);
 
-int __libc_recvmsg(int a, struct msghdr* msg, int flags);
-int __libc_recvmsg(int a, struct msghdr* msg, int flags) {
+ssize_t __libc_recvmsg(int a, struct msghdr* msg, int flags);
+ssize_t __libc_recvmsg(int a, struct msghdr* msg, int flags) {
   long args[] = { a, (long) msg, flags };
   return socketcall(SYS_RECVMSG, args);
 }
 
-int recvmsg(int a, struct msghdr *msg, int flags)
+ssize_t recvmsg(int a, struct msghdr *msg, int flags)
  __attribute__ ((weak,alias("__libc_recvmsg"))) ;
 
 #endif
